NPC stats edits by Angry Catster
Yet another update for npc files:

Even more npcs are covered

Level and aggro range stats are also corrected now

Newly implemented stats are also added to npcs. As I said before, new stats will not be noticeable most of the time. However, this makes npcs fell a tiny bit closer to retail and (more importantly) gives more opportunities for tweaking and custom npc creation.

In comparison to the current github source, the following npc stats are now working:

main_hand_accuracy – npc accuracy stat used against target’s evasion/block/parry stats. An npc with 3000 accuracy will never miss against a player with 2500 evasion, but will miss 30% of the time against the player with 3300 evasion.

main_hand_crit_rate – affects chance of an npc performing a critical hit

mdef – reduces the spell power of the attacker. In simpler words, more mdef – less magic spell damage taken by npc

mresist – gives a chance to outright ignore effects of the spell used against an npc

evasion – If it is higher than target’s accuracy stat, then an npc will be able to dodge their attacks

block - If it is higher than target’s accuracy stat, then an npc will be able to block their attacks

parry - If it is higher than target’s accuracy stat, then an npc will be able to parry their attacks

All of that required some changes to my spaghetti program, so some playtesting is needed to make sure that everyone has the right stats (spent a couple of hours running around and checking stats in game and everything seems to be right). Attached zip has all npc-related changes I made so far – no need to take files from my previous posts – jut put these in your source and compile it.

 - Angry Catster

+++ src/com/aionemu/gameserver/model/templates/stats/NpcStatsTemplate.java
@@ -1,5 +1,5 @@
 /*
-
+ * This file is part of Encom. **ENCOM FUCK OTHER SVN**
  *
  *  Encom is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Lesser Public License as published by
@@ -47,7 +47,7 @@ public class NpcStatsTemplate extends StatsTemplate
 	private int power;
 	
 	@XmlAttribute(name = "maxXp")
-	private int maxXp;
+	private long maxXp;
 	
 	public float getRunSpeedFight() {
 		return runSpeedFight;
@@ -57,7 +57,7 @@ public class NpcStatsTemplate extends StatsTemplate
 		return pdef;
 	}
 	
-	public float getMdef() {
+	public int getMdef() {
 		return mdef;
 	}
 	
@@ -81,7 +81,7 @@ public class NpcStatsTemplate extends StatsTemplate
 		this.power = power;
 	}
 	
-	public int getMaxXp() {
+	public long getMaxXp() {
 		return maxXp;
 	}
 }
